requirements.txt
# ASAN Macro - Maritime & Trade Analysis Agent
# Python 3.9+

# LLM
openai>=1.0.0
google-generativeai>=0.3.0

# Database & data
# (sqlite3 is in stdlib)

# Env and CLI
python-dotenv>=1.0.0

# Optional: dynamic data ingestion (CSV from URL, JSON API)
requests>=2.28.0

# Optional: UN Comtrade API (for scripts/fetch_comtrade.py live data)
comtradeapicall>=1.2.0
pandas>=1.0.0
